TINY-6413: Fixed a number of incorrect or hard to use types found while typing other projects
