+++ modules/tinymce/src/core/main/ts/api/Formatter.ts
@@ -9,13 +9,13 @@ import { Cell, Fun } from '@ephox/katamari';
 import * as CaretFormat from '../fmt/CaretFormat';
 import * as FormatChanged from '../fmt/FormatChanged';
 import { FormatRegistry } from '../fmt/FormatRegistry';
+import { Format, FormatVars } from '../fmt/FormatTypes';
 import * as MatchFormat from '../fmt/MatchFormat';
 import * as Preview from '../fmt/Preview';
 import * as FormatShortcuts from '../keyboard/FormatShortcuts';
 import * as Rtc from '../Rtc';
 import { RangeLikeObject } from '../selection/RangeTypes';
 import Editor from './Editor';
-import { Format, FormatVars } from './fmt/Format';
 
 /**
  * Text formatter engine class. This class is used to apply formats like bold, italic, font size
