+++ modules/tinymce/src/core/main/ts/api/fmt/StyleFormat.ts
